handle error in create method, change param id to int64, and implement method find by id

+++ internal/seller/usecases/seller_usecase.go
@@ -5,7 +5,7 @@ import model "Dzaakk/simple-commerce/internal/seller/models"
 type SellerUseCase interface {
 	Create(data model.ReqCreate) (int64, error)
 	Update(data model.ReqUpdate) (int64, error)
-	FindById(sellerId int) (*model.ResData, error)
+	FindById(sellerId int64) (*model.ResData, error)
 	FindByUsername(username string) (*model.ResData, error)
 	Deactivate(sellerId int) (int64, error)
 }
